[BM-664] Prepare and refactor packages and naming (#241)
* Add functionality to change voice analysis option from ML to noise detect

Parent device can change voiceAnalysisOption, this action is sent to
baby device via websockets and handled there. Noise detection is done
by simple recording data analysis which give us values that should be
close to sound decibels ^^.

* Migration fix - column constrains needs to stay the same (nullable)

* Fallback to destructive migration - just in case.

* Kotlin update, android gradle plugin update

* Changed packages and DI modules to better reflect actual project setup

* Return to old formatting
