+++ app/src/main/kotlin/co/netguru/baby/monitor/client/application/di/BabyMonitorGlideModule.kt
@@ -1,4 +1,4 @@
-package co.netguru.baby.monitor.client.application
+package co.netguru.baby.monitor.client.application.di
 
 import com.bumptech.glide.annotation.GlideModule
 import com.bumptech.glide.module.AppGlideModule
