+++ app/src/main/kotlin/co/netguru/baby/monitor/client/application/firebase/FirebaseRepository.kt
@@ -5,7 +5,7 @@ import android.net.Uri
 import co.netguru.baby.monitor.client.R
 import co.netguru.baby.monitor.client.application.App
 import co.netguru.baby.monitor.client.common.RunsInBackground
-import co.netguru.baby.monitor.client.feature.machinelearning.WavFileGenerator
+import co.netguru.baby.monitor.client.feature.voiceAnalysis.WavFileGenerator
 import com.google.firebase.FirebaseApp
 import com.google.firebase.storage.FirebaseStorage
 import com.google.firebase.storage.StorageMetadata
