+++ app/src/main/kotlin/co/netguru/baby/monitor/client/feature/firebasenotification/NotificationType.kt
@@ -3,5 +3,6 @@ package co.netguru.baby.monitor.client.feature.firebasenotification
 enum class NotificationType {
     CRY_NOTIFICATION,
     LOW_BATTERY_NOTIFICATION,
+    NOISE_NOTIFICATION,
     DEFAULT
 }
