+++ app/src/main/kotlin/co/netguru/baby/monitor/client/feature/voiceAnalysis/VoiceAnalysisOption.kt
@@ -1,4 +1,4 @@
-package co.netguru.baby.monitor.client.feature.machinelearning
+package co.netguru.baby.monitor.client.feature.voiceAnalysis
 
 enum class VoiceAnalysisOption {
     MachineLearning,
