+++ build.gradle
@@ -2,7 +2,7 @@
 buildscript {
 
     ext {
-        kotlinVersion = '1.3.60'
+        kotlinVersion = '1.3.61'
     }
     repositories {
         google()
@@ -14,7 +14,7 @@ buildscript {
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.5.2'
+        classpath 'com.android.tools.build:gradle:3.5.3'
         classpath "com.getkeepsafe.dexcount:dexcount-gradle-plugin:0.8.6"
         classpath 'com.vanniktech:gradle-android-junit-jacoco-plugin:0.15.0'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlinVersion"
